Better roof momentary button control
Removes [Stop!] button if no way to stop the roof is specified.
Shows a [Press!] button instead of [Stop!] if the only way to stop the roof is by pressing a single Open/Close/Stop button.
In these modes roof errors that result in the roof needing to be stopped do NOT trigger a stop press since the effect of pressing the single button is not known with 100% confidence.